requirements.txt
anyio==3.3.4
appnope==0.1.2
argon2-cffi==20.1.0
async-generator==1.10
attrs==21.2.0
Babel==2.9.1
backcall==0.2.0
bcrypt==3.2.0
bleach==3.3.0
cachetools==4.2.2
camelot-py==0.8.2
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==8.0.1
coverage==5.5
cramjam==2.3.2
cryptography==3.4.7
cv==1.0.0
cycler==0.10.0
decorator==5.0.9
defusedxml==0.7.1
distro==1.5.0
entrypoints==0.3
et-xmlfile==1.1.0
fastparquet==0.6.3
Flask==2.0.0
fsspec==2021.6.1
google-api-core==1.28.0
google-api-python-client==2.36.0
google-auth==1.30.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
google-cloud-core==1.6.0
google-cloud-documentai==0.5.0
google-cloud-secret-manager==2.8.0
google-cloud-storage==1.40.0
google-cloud-translate==3.2.1
google-crc32c==1.1.2
google-resumable-media==1.3.0
googleapis-common-protos==1.53.0
grpc-google-iam-v1==0.12.3
grpcio==1.38.0
gunicorn==20.1.0
httplib2==0.20.2
idna==2.10
importlib-metadata==4.0.1
ipycanvas==0.8.2
ipykernel==5.5.5
ipython==7.23.1
ipython-genutils==0.2.0
ipywidgets==7.6.3
itsdangerous==2.0.1
jedi==0.18.0
Jinja2==3.0.1
joblib==1.0.1
json5==0.9.6
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.12
jupyter-console==6.4.0
jupyter-core==4.7.1
jupyter-server==1.11.1
jupyterlab==3.2.1
jupyterlab-pygments==0.1.2
jupyterlab-server==2.8.2
jupyterlab-widgets==1.0.0
kiwisolver==1.3.1
Levenshtein==0.16.0
libcst==0.4.0
lxml==4.6.3
MarkupSafe==2.0.1
matplotlib==3.4.2
matplotlib-inline==0.1.2
mistune==0.8.4
mypy-extensions==0.4.3
nbclassic==0.3.2
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.4.0
numpy==1.20.3
oauth2client==4.1.3
oauthlib==3.1.1
opencv-python==4.5.2.52
openpyxl==3.0.7
orjson==3.5.2
packaging==20.9
pandas==1.2.4
pandocfilters==1.4.3
paramiko==2.8.1
parso==0.8.2
pdf2image==1.15.1
pdfminer.six==20201018
pexpect==4.8.0
pickleshare==0.7.5
pikepdf==2.12.1
Pillow==8.2.0
prometheus-client==0.10.1
prompt-toolkit==3.0.18
proto-plus==1.18.1
protobuf==3.17.0
ptyprocess==0.7.0
pyarrow==4.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyDrive==1.3.1
Pygments==2.9.0
PyNaCl==1.4.0
pyOpenSSL==20.0.1
pyparsing==2.4.7
PyPDF2==1.26.0
pyperclip==1.8.2
pyrsistent==0.17.3
pysftp==0.2.9
pytesseract==0.3.7
python-dateutil==2.8.1
python-docx==0.8.11
python-dotenv==0.19.0
pytz==2021.1
PyYAML==5.4.1
pyzmq==22.0.3
qtconsole==5.1.0
QtPy==1.9.0
rapidfuzz==1.8.2
requests==2.25.1
requests-oauthlib==1.3.0
requests-unixsocket==0.2.0
rsa==4.7.2
scikit-learn==0.24.2
scipy==1.6.3
Send2Trash==1.5.0
Shapely==1.7.1
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
tabula-py==2.2.0
terminado==0.10.0
testpath==0.5.0
threadpoolctl==2.1.0
thrift==0.13.0
tika==1.24
tornado==6.1
tqdm==4.61.1
traitlets==5.0.5
typing-extensions==3.10.0.0
typing-inspect==0.7.1
uritemplate==4.1.1
urllib3==1.26.4
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.2.1
Werkzeug==2.0.1
widgetsnbextension==3.5.1
WTForms==2.3.3
zipp==3.4.1
